requirement.txt
Flask==2.3.2
opencv-python==4.6.0.66
python-dotenv==1.0.0
ultralytics==8.0.14
